+++ Changelog.md
@@ -9,3 +9,9 @@
 # 1.0.1 Release
 
 * Just a name and icon change.
+
+# 1.0.2 Release
+
+* New look for the launcher menu! (By Devvy)
+
+* New work in progress login system! (By Devvy)